Stop caching the run task
Docker and the CRaC JDK may have changed since the last run.

Previously, we were caching the result of running the tests, which meant it was always from cache even it Docker or the CRaC JDK had updated.

This change stops caching the TestScriptRunnerTask, so we should run the test script every time this action is scheduled.