Add requirements for tests
Some tests will require services such as MySQL or Redis.
To enable this, add a list of service strings to the metadata.
Then if these requirements exist, call functions of the same name to start and stop the containers.

To enable all this, we create a docker network by default on startup.And delete it at the end